benchmarks - spectral-norm cast to DOUBLE to reroute to math16 _Float16

+++ support/benchmarks/spectral-norm/z88dk-classic/spectral-norm.c
@@ -48,7 +48,8 @@
 
 DOUBLE eval_A(int i, int j)
 {
-    return 1.0/((i+j)*(i+j+1)/2+i+1);
+    /* Cast so sccz80 uses f16 under --math16; bare 1.0 is f48 (ddiv). */
+    return (DOUBLE)1.0/(DOUBLE)((i+j)*(i+j+1)/2+i+1);
 }
 
 void eval_A_times_u(const DOUBLE u[], DOUBLE Au[])
